+++ Integrations/rasterize/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
   - Added the `rasterize-pdf` command, which converts a PDF file to an image file.
   - `rasterize-email` now available in offline mode. 
+  - New `wait_time` parameter in `rasterize` command that sets the time to wait before taking the screenshot
   
 ## [19.10.1] - 2019-10-15
   - Updated to use Chrome driver instead of phantomJS (requires Demisto 5.0).